+++ src/kikuchipy/signals/_kikuchipy_signal.py
@@ -26,7 +26,6 @@ from typing import Any
 import warnings
 
 import dask.array as da
-from hyperspy._lazy_signals import LazySignal2D
 from hyperspy.signals import Signal2D
 import numpy as np
 from packaging.version import Version
@@ -34,7 +33,7 @@ from skimage.util.dtype import dtype_range
 import yaml
 
 from kikuchipy._constants import dependency_version
-from kikuchipy._utils.rosettasciio import RGB_DTYPES
+from kikuchipy._utils.rosettasciio_utils import RGB_DTYPES
 from kikuchipy.pattern._pattern import (
     _adaptive_histogram_equalization,
     normalize_intensity,
@@ -43,6 +42,11 @@ from kikuchipy.pattern._pattern import (
 from kikuchipy.signals.util._overwrite_hyperspy_methods import (
     insert_doc_disclaimer,
 )
+
+if dependency_version["hyperspy"] >= Version("2.4.0"):
+    from hyperspy.signals import LazySignal2D
+else:
+    from hyperspy._lazy_signals import LazySignal2D
 
 _logger = logging.getLogger(__name__)
 
@@ -296,7 +300,7 @@ class KikuchipySignal2D(Signal2D):
         >>> np.mean(s.data)
         np.float64(146.0670987654321)
         >>> s.normalize_intensity(dtype_out=np.float32)
-        >>> np.mean(s.data)
+        >>> np.mean(s.data)  # doctest: +SKIP
         np.float32(0.0)
         """
         if lazy_output and inplace: