+++ src/kikuchipy/signals/ebsd.py
@@ -44,7 +44,7 @@ from scipy.ndimage import correlate, gaussian_filter
 from skimage.util.dtype import dtype_range
 
 from kikuchipy._constants import dependency_version, pyopencl_context_available
-from kikuchipy.detectors.ebsd_detector import EBSDDetector
+from kikuchipy.detectors._ebsd_detector import EBSDDetector
 from kikuchipy.filters.fft_barnes import _fft_filter, _fft_filter_setup
 from kikuchipy.filters.window import Window
 from kikuchipy.indexing._dictionary_indexing import _dictionary_indexing
@@ -157,7 +157,15 @@ class EBSD(KikuchipySignal2D):
     >>> s
     <EBSD, title: patterns Scan 1, dimensions: (3, 3|60, 60)>
     >>> s.detector
-    EBSDDetector(shape=(60, 60), pc=(0.425, 0.213, 0.501), sample_tilt=70.0, tilt=0.0, azimuthal=0.0, twist=0.0, binning=8.0, px_size=1.0 um)
+    EBSDDetector
+      shape (Ny, Nx):     (60, 60)
+      pc (PCx, PCy, PCz): (0.425, 0.213, 0.501)
+      sample_tilt:        70.0 deg
+      tilt:               0.0 deg
+      azimuthal:          0.0 deg
+      twist:              0.0 deg
+      binning:            8
+      px_size:            1.0 um
     >>> s.static_background
     array([[84, 87, 90, ..., 27, 29, 30],
            [87, 90, 93, ..., 27, 28, 30],
@@ -165,7 +173,7 @@ class EBSD(KikuchipySignal2D):
            ...,
            [80, 82, 84, ..., 36, 30, 26],
            [79, 80, 82, ..., 28, 26, 26],
-           [76, 78, 80, ..., 26, 26, 25]], dtype=uint8)
+           [76, 78, 80, ..., 26, 26, 25]], shape=(60, 60), dtype=uint8)
     >>> s.xmap
     Phase  Orientations  Name  Space group  Point group  Proper point group     Color
         0    9 (100.0%)    ni        Fm-3m         m-3m                 432  tab:blue
@@ -494,7 +502,7 @@ class EBSD(KikuchipySignal2D):
                ...,
                [80, 82, 84, ..., 36, 30, 26],
                [79, 80, 82, ..., 28, 26, 26],
-               [76, 78, 80, ..., 26, 26, 25]], dtype=uint8)
+               [76, 78, 80, ..., 26, 26, 25]], shape=(60, 60), dtype=uint8)
 
         The static background can be removed by subtracting or dividing
         this background from each pattern:
@@ -1155,7 +1163,7 @@ class EBSD(KikuchipySignal2D):
             raise ValueError("'lazy_output=True' requires 'inplace=False'")
 
         if not isinstance(factor, int) or factor <= 1:
-            raise ValueError(f"Binning `factor` {factor} must be an integer > 1")
+            raise ValueError(f"Binning factor {factor} must be an integer > 1")
         else:
             factor = int(factor)
 
@@ -1163,11 +1171,12 @@ class EBSD(KikuchipySignal2D):
         rest = np.mod(sig_shape_old, factor)
         if not all(rest == 0):
             raise ValueError(
-                f"Binning `factor` {factor} must be a divisor of the initial pattern "
+                f"Binning factor {factor} must be a divisor of the initial pattern "
                 f"shape {sig_shape_old}, but {tuple(rest)} pixels remain.\n"
-                "You might try to crop away these pixels first using EBSD.crop()"
+                "You might try to crop away these pixels first using EBSD.crop()."
             )
-        sig_shape_new = tuple(np.array(sig_shape_old) // factor)
+        sig_shape_new = np.array(sig_shape_old) // factor
+        sig_shape_new = tuple(map(int, sig_shape_new))
         sig_shape_new = sig_shape_new[::-1]
 
         if dtype_out is not None:
@@ -1185,8 +1194,9 @@ class EBSD(KikuchipySignal2D):
             static_bg_new = _downsample2d(static_bg, factor, omin, omax, dtype_out)
             attrs["static_background"] = static_bg_new
 
-        attrs["detector"].shape = sig_shape_new
-        attrs["detector"].binning *= factor
+        detector: EBSDDetector = attrs["detector"]
+        detector.shape = sig_shape_new
+        detector.binning *= factor
 
         map_kw = {
             "show_progressbar": show_progressbar,
@@ -2803,17 +2813,18 @@ class EBSD(KikuchipySignal2D):
             attrs["static_background"] = static_bg2
 
         # Update detector shape and binning factor
-        attrs["detector"].shape = sig_shape_new
+        detector: EBSDDetector = attrs["detector"]
+        detector.shape = sig_shape_new
         factors = np.array(sig_shape_old) / np.array(sig_shape_new)
-        binning = attrs["detector"].binning * factors
-        if binning[0] == binning[1] and np.allclose(binning, binning.round(0)):
-            attrs["detector"].binning = int(binning[0])
+        binning2 = detector.binning * factors
+        if binning2[0] == binning2[1] and np.allclose(binning2, binning2.round(0)):
+            detector.binning = float(binning2[0])
         else:
-            attrs["detector"].binning = 1
+            detector.binning = 1.0
 
         if nav_shape_new != nav_shape_old:
             attrs["xmap"] = None
-            attrs["detector"].pc = np.full(nav_shape_new + (3,), 0.5)
+            detector.pc = np.full(nav_shape_new + (3,), 0.5)
             attrs["static_background"] = None
 
         new._set_custom_attributes(attrs)
@@ -3182,9 +3193,9 @@ class LazyEBSD(LazyKikuchipySignal2D, EBSD):
     See the documentation of ``EBSD`` for attributes and methods.
 
     This class extends HyperSpy's
-    :class:`~hyperspy._signals.signal2d.LazySignal2D` class for EBSD
-    patterns. See the documentation of that class for how to create
-    this signal and the list of inherited attributes and methods.
+    :class:`~hyperspy.signals.LazySignal2D` class for EBSD patterns. See
+    the documentation of that class for how to create this signal and
+    the list of inherited attributes and methods.
     """
 
     def compute(self, *args, **kwargs) -> None:
@@ -3289,7 +3300,7 @@ def _update_custom_attributes(
     nav_slices: slice | tuple | None = None,
     sig_slices: slice | tuple | None = None,
     new_nav_shape: tuple[int, ...] | None = None,
-    new_sig_shape: tuple[int, ...] | None = None,
+    new_sig_shape: tuple[int, int] | None = None,
 ) -> dict:
     """Update dictionary of custom attributes after slicing the signal
     data.